requirements.txt
absl-py==0.13.0
alabaster==0.7.12
appdirs==1.4.4
asgiref==3.4.1
astunparse==1.6.3
autopep8==1.5.7
beautifulsoup4==4.10.0
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.6
charset-normalizer==2.0.4
clang==5.0
cryptography==3.4.7
cycler==0.10.0
decouple==0.0.7
defusedxml==0.7.1
distlib==0.3.1
dj-database-url==0.5.0
Django==3.2.6
django-admin-interface==0.16.3
django-allauth==0.45.0
django-bootstrap-v5==1.0.5
django-bootstrap4==3.0.1
django-colorfield==0.4.2
django-filter==2.4.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-rest-auth==0.9.5
django-user-agents==0.4.0
djangorestframework==3.12.4
djangorestframework-jsonapi==4.2.1
djangorestframework-simplejwt==4.8.0
docopt==0.6.2
filelock==3.0.12
flatbuffers==1.12
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.40.0
gunicorn==20.1.0
h5py==3.1.0
idna==3.2
imageai==2.1.5
inflection==0.5.1
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
Markdown==3.3.4
matplotlib==3.4.3
mysqlclient==2.0.3
numpy==1.19.5
oauthlib==3.1.1
opencv-contrib-python==4.5.3.56
opencv-python==4.5.3.56
opt-einsum==3.3.0
Pillow==8.3.2
pipenv==2020.11.15
pipreqs==0.4.10
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
PyJWT==2.1.0
PyMySQL==1.0.2
pyparsing==2.4.7
python-dateutil==2.8.2
python-decouple==3.4
python-wordpress-xmlrpc==1.4
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.7.1
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.2
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.0
tensorflow-estimator==2.6.0
termcolor==1.1.0
toml==0.10.2
typing-extensions==3.7.4.3
ua-parser==0.10.0
Unipath==1.1
urllib3==1.26.6
user-agents==2.2.0
virtualenv==20.4.4
virtualenv-clone==0.5.4
Werkzeug==2.0.1
whitenoise==5.3.0
wrapt==1.12.1
yarg==0.1.9
